+++ packages/vscode-extension/src/debugging/DebugAdapter.ts
@@ -30,7 +30,7 @@ import { AnyObject } from "vscode-js-debug/src/adapter/objectPreview/betterTypes
 import { messageFormatters, previewAsObject, previewRemoteObject } from "vscode-js-debug/src/adapter/objectPreview";
 import { formatMessage as externalFormatMessage } from "vscode-js-debug/src/adapter/messageFormat";
 import { PreviewContextType } from "vscode-js-debug/src/adapter/objectPreview/contexts";
-import { getSourceSuffix } from "vscode-js-debug/src/adapter/templates";
+import getArraySlots from "./templates";
 
 function typeToCategory(type: string) {
   switch (type) {
@@ -149,8 +149,8 @@ export class DebugAdapter extends DebugSession {
     }
   };
 
+  // Based on https://github.com/microsoft/vscode-js-debug/blob/3be255753c458f231e32c9ef5c60090236780060/src/adapter/console/textualMessage.ts#L83
   async formatDefaultString (
-    // thread: Thread,
     args: ReadonlyArray<Cdp.Runtime.RemoteObject>,
   ) {
     const useMessageFormat = args.length > 1 && args[0].type === 'string';
@@ -163,50 +163,13 @@ export class DebugAdapter extends DebugSession {
     if (formatResult.usedAllSubs && !args.some(previewAsObject)) {
       return { output };
     } else {
-      return await this.formatComplexStringOutput(output, args);
+      const outputVar = await this.createVariableForOutputEvent(args as CDPRemoteObject[]);
+
+      return { output, variablesReference: outputVar };
     }
   };
 
-  async formatComplexStringOutput (
-    // thread: Thread,
-    output: string,
-    args: ReadonlyArray<Cdp.Runtime.RemoteObject>,
-  ) {
-    const outputVar = await this.createVariableForOutputEvent(args as CDPRemoteObject[]);
-
-    return { output, variablesReference: outputVar };
-  };
-
-
   private async handleConsoleAPICall(message: any) {
-    const string = 'string';
-    const object = { key: 'value', key2: 'value2' };
-    const set = new Set([1, 2, 3, 3]);
-    const array100 = new Array(100).fill(0).map((_, i) => i);
-    const array1000 = new Array(1000).fill(0).map((_, i) => i);
-    const int = 1;
-    const float = 1.1;
-    const func = function() {};
-    const anonymousFunc = () => {};
-
-
-    // console.log(string);
-    // console.log(object);
-    // console.log(set);
-    // console.log(array100);
-    // console.log(array1000);
-    // console.log(int);
-    // console.log(float);
-    // console.log(func);
-    // console.log(anonymousFunc);
-    // console.log(string, object, set,
-    //   array100,
-    //   array1000,
-    //   int,
-    //   float,
-    //   func,
-    //   anonymousFunc);
-
     // We wrap console calls and add stack information as last three arguments, however
     // some logs may baypass that, especially when printed in initialization phase, so we
     // need to detect whether the wrapper has added the stack info or not
@@ -238,26 +201,9 @@ export class DebugAdapter extends DebugSession {
       const testOutput = await this.formatDefaultString(message.params.args.slice(0, -3));
 
       output = new OutputEvent(
-        // (await formatMessage(message.params.args.slice(0, -3))) + "\n",
         testOutput.output,
         typeToCategory(message.params.type)
       );
-      
-      // Old method to compare
-      // output = new OutputEvent(
-      //   (await formatMessage(message.params.args.slice(0, -3))) + "\n",
-      //   typeToCategory(message.params.type)
-      // );
-      // output.body = {
-      //   ...output.body,
-      //   //@ts-ignore source, line, column and group are valid fields
-      //   source: new Source(sourceURL, sourceURL),
-      //   line: this.linesStartAt1 ? lineNumber1Based : lineNumber1Based - 1,
-      //   column: this.columnsStartAt1 ? columnNumber0Based + 1 : columnNumber0Based,
-      //   variablesReference: variablesRefDapID,
-      // };   
-      
-      // this.sendEvent(output);
       
       output.body = {
         ...testOutput,
@@ -289,55 +235,6 @@ export class DebugAdapter extends DebugSession {
   private async createVariableForOutputEvent(args: CDPRemoteObject[]) {
     const prepareVariables = await Promise.all(
       args.map(async (arg: CDPRemoteObject, index: number) => {
-        // if (arg.type === "object") {
-        //   if (arg?.subtype === "array") {
-        //     const match = String(arg.description).match(/\(([0-9]+)\)/);
-        //     const length = match ? +match[1] : 0;
-
-        //     // if (length < 100) {
-        //     //   arg.description = previewRemoteObject(arg, PreviewContextType.PropertyValue);
-        //     //   arg.objectId = this.variableStore.adaptCDPObjectId(arg.objectId).toString();
-
-        //     //   return [{ name: `arg${index}`, value: arg }];
-        //     // }
-
-        //     // const stackObjectProperties = await this.variableStore.getProperties(
-        //     //   this.variableStore.adaptCDPObjectId(arg.objectId),
-        //     //   (params: object) => {
-        //     //     const response = this.cdpSession.sendCDPMessage("Runtime.getProperties", params);
-        //     //     return response;
-        //     //   }
-        //     // );
-
-        //     // console.log("LOG PROPERTIES -> ", stackObjectProperties);
-
-        //     // const chunkSize = 100;
-        //     // const chunks = [];
-        //     // for (let i = 0; i < length; i += chunkSize) {
-        //     //     const chunk = stackObjectProperties.slice(i, i + chunkSize);
-        //     //     // do whatever
-        //     //     const chunkVariablesObjectDapID = this.variableStore.pushReplVariable(chunk.map(value => ({
-        //     //       value: value.value,
-        //     //       name: value.name,
-        //     //     })));
-                
-        //     //     chunks.push({
-        //     //       name: `[${i}..${i + chunkSize - 1}]`,
-        //     //       value: {
-        //     //         type: "object",
-        //     //         objectId: chunkVariablesObjectDapID.toString(),
-        //     //         description: '',
-        //     //       },
-        //     //     });
-        //     // }
-            
-        //     arg.description = previewRemoteObject(arg, PreviewContextType.PropertyValue);
-        //     arg.objectId = this.variableStore.adaptCDPObjectId(arg.objectId).toString();
-        //     return { name: `arg${index}`, value: arg, indexedVariables: length > 100 ? length : undefined,
-        //     namedVariables: length > 100 ? 1 : undefined};
-        //   }
-        // }
-
         if (arg.type === "object") {
           arg.description = previewRemoteObject(arg, PreviewContextType.PropertyValue);
           arg.objectId = this.variableStore.adaptCDPObjectId(arg.objectId).toString();
@@ -543,41 +440,9 @@ export class DebugAdapter extends DebugSession {
     response.body = response.body || {};
     response.body.variables = [];
     
-    if (args.filter !== "indexed" && args.filter !== "named") {
-      response.body.variables = await this.variableStore.get(
-        args.variablesReference,
-        (params: object) => {
-          return this.cdpSession.sendCDPMessage("Runtime.getProperties", params);
-        }
-      );
-    }
-
-    console.log('args', args);
-    console.log('response', response);
-
     if (args.filter === "indexed") {
-      const fn = function(
-        this: unknown[],
-        start: number,
-        count: number,
-      ) {
-        const result = {};
-        const from = start === -1 ? 0 : start;
-        const to = count === -1 ? this.length : start + count;
-        for (let i = from; i < to && i < this.length; ++i) {
-          const descriptor = Object.getOwnPropertyDescriptor(this, i);
-          if (descriptor) {
-            Object.defineProperty(result, i, descriptor);
-          }
-        }
+      const stringified = '' + getArraySlots;
 
-        return result;
-      };      
-
-      let stringified = '' + fn;
-      // const endIndex = stringified.lastIndexOf('}');
-      // stringified = stringified.slice(0, endIndex) + getSourceSuffix(undefined)
-      //   + stringified.slice(endIndex);
       try {
         const partialValue = await this.cdpSession.sendCDPMessage("Runtime.callFunctionOn", {
           functionDeclaration: stringified,
@@ -585,25 +450,26 @@ export class DebugAdapter extends DebugSession {
           arguments: [args.start, args.count].map(value => ({ value })),
         });
 
-        console.log('partialValue', partialValue);
-        const cdpValue = this.variableStore.adaptCDPObjectId(partialValue.result.objectId);
-
         const properties = await this.variableStore.get(
-          cdpValue,
+          this.variableStore.adaptCDPObjectId(partialValue.result.objectId),
           (params: object) => {
             return this.cdpSession.sendCDPMessage("Runtime.getProperties", params);
           }
         );
 
-
         response.body.variables = properties;
-
-        this.sendResponse(response);
-
-        return;
       } catch(e) {
-        console.log('error', e);
+        Logger.error('[CDP] Failed to retrieve array partially', e);
       }
+    }
+
+    if (args.filter !== "indexed" && args.filter !== "named") {
+      response.body.variables = await this.variableStore.get(
+        args.variablesReference,
+        (params: object) => {
+          return this.cdpSession.sendCDPMessage("Runtime.getProperties", params);
+        }
+      );
     }
