+++ packages/vscode-extension/src/debugging/templates/index.ts
@@ -1,0 +1,1 @@
+export { default } from'./getArraySlots';
